lukes amazing code/requirements.txt
requests>=2.31.0
tweepy>=4.14.0
python-dotenv>=1.0.0 